+++ src/main/java/turtleMart/order/dto/request/CartOrderSheetRequest.java
@@ -8,6 +8,6 @@ public record CartOrderSheetRequest(
         Long productId,
         @NotNull(message = "수량은 필수입니다.")
         @Min(value = 1, message = "수량은 1 이상이어야 합니다.")
-        Long quantity
+        Integer quantity
 ) {
 }
